Move reps_cache to nano/secure
This is needed, because reps_cache will get access to the new rep_cache
data store.
