+++ nano/secure/CMakeLists.txt
@@ -53,6 +53,8 @@ add_library(
   network_filter.cpp
   pending_info.hpp
   pending_info.cpp
+  rep_weights.hpp
+  rep_weights.cpp
   utility.hpp
   utility.cpp
   vote.hpp
