+++ nano/secure/rep_weights.cpp
@@ -1,4 +1,4 @@
-#include <nano/lib/rep_weights.hpp>
+#include <nano/secure/rep_weights.hpp>
 #include <nano/store/component.hpp>
 
 void nano::rep_weights::representation_add (nano::account const & source_rep_a, nano::uint128_t const & amount_a)
